# swagger/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "varsion": "0.0.1",
    "title": "DAS SDL API server",
    "description": "DAS SDL 서버 API swagger 문서",
    "version": "1.0.0"
  },
  "host": "localhost:3200",
  "basePath": "/das",
...
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "token",
      "description": "유저token"
    }
  },
  "paths": {
...
    "/prj/ctPrj": {
      "post": {
        "tags": [
          "prj"
        ],
        "summary": "프로젝트 생성",
        "description": "사용자가 프로젝트를 생성한다.",
        "consumes": ["multipart/form-data"],
        "parameters": [
          {
            "name": "prj_name",
            "in": "formData",
            "type": "string",
            "required": true,
            "example": "프로젝트1",
            "description" : "프로젝트 이름"
          },
          {
            "name": "prj_description",
            "in": "formData",
            "type": "string",
            "required": true,
            "example": "이 프로젝트는 러시아에서 시작되어 전세계적으로 펼쳐져..",
            "description" : "프로젝트 설명"
          },
          {
            "name": "prj_start_version",
            "in": "formData",
            "type": "string",
            "required": true,
            "example": "v0.0.1",
            "description" : "프로젝트 버전"
          },
          {
            "name": "rgst_user_id",
            "in": "formData",
            "type": "number",
            "required": true,
            "example": 1,
            "description" : "프로젝트 생성 유저 아이디"
          },
          {
            "name": "prj_lnk",
            "in": "formData",
            "type": "string", 
            "example": "http://example.com",
            "required" : true,
            "description" : "프로젝트 링크"
          },
          {
            "name": "prj_sec_user",
            "in": "formData",
            "type": "array",
            "items" : {
              "type" : "number"
            },
            "required": true,
            "example": 2,
            "description" : "프로젝트 보안 담당자(다중)"
          },
          {
            "name": "prj_dev_user",
            "in": "formData",
            "type": "array",
            "items" : {
              "type" : "number"
            },
            "required": true,
            "example": 1,
            "description" : "프로젝트 개발 담당자(다중)"
          },
          {
            "name": "files",
            "in": "formData",
            "type": "array",
            "items" : {
              "type" : "file"
            },
            "required": true,
            "description" : "프로젝트 파일(다중)"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "resultCode": {
                  "type": "number",
                  "example": 200
                },
                "resultMsg": {
                  "type": "string",
                  "example": "프로젝트 생성 완료"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        },
        "security": [
          {
            "apiKeyAuth": [
              {
                "token": ""
              }
            ]
          }
        ]
      }
    },
    "/prj/addPrjVer": {
      "post": {
        "tags": [
          "prj"
        ],
        "summary": "프로젝트 버전 생성",
        "description": "사용자가 프로젝트 내 버전을 생성한다.",
        "consumes": ["multipart/form-data"],
        "parameters": [
          {
            "name": "prj_id",
            "in": "formData",
            "type": "number",
            "required": true,
            "example": "1",
            "description" : "프로젝트 아이디"
          },
          {
            "name": "prj_name",
            "in": "formData",
            "type": "string",
            "required": true,
            "example": "프로젝트1",
            "description" : "프로젝트 이름"
          },
          {
            "name": "prj_description",
            "in": "formData",
            "type": "string",
            "required": true,
            "example": "이 프로젝트는 러시아에서 시작되어 전세계적으로 펼쳐져..",
            "description" : "프로젝트 설명"
          },
          {
            "name": "version_number",
            "in": "formData",
            "type": "string",
            "required": true,
            "example": "v0.0.1",
            "description" : "프로젝트 버전"
          },
          {
            "name": "rgst_user_id",
            "in": "formData",
            "type": "number",
            "required": true,
            "example": 1,
            "description" : "프로젝트 버전 생성 유저 아이디"
          },
          {
            "name": "prj_lnk",
            "in": "formData",
            "type": "string", 
            "example": "http://example.com",
            "required" : true,
            "description" : "프로젝트 링크"
          },
          {
            "name": "prj_sec_user",
            "in": "formData",
            "type": "array",
            "items" : {
              "type" : "number"
            },
            "required": true,
            "example": 2,
            "description" : "프로젝트 보안 담당자(다중)"
          },
          {
            "name": "prj_dev_user",
            "in": "formData",
            "type": "array",
            "items" : {
              "type" : "number"
            },
            "required": true,
            "example": 1,
            "description" : "프로젝트 개발 담당자(다중)"
          },
          {
            "name": "files",
            "in": "formData",
            "type": "array",
            "items" : {
              "type" : "file"
            },
            "required": true,
            "description" : "프로젝트 파일(다중)"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "resultCode": {
                  "type": "number",
                  "example": 200
                },
                "resultMsg": {
                  "type": "string",
                  "example": "프로젝트 버전 생성 완료"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        },
        "security": [
          {
            "apiKeyAuth": [
              {
                "token": ""
              }
            ]
          }
        ]
      }
    },
    "/prj/prjList": {
      "get": {
        "tags": [
          "prj"
        ],
        "summary": "프로젝트 조회",
        "description": "사용자가 프로젝트를 조회한다.",
        "parameters" : [
          {
            "name" : "user_id",
            "description" : "유저 아이디 값",
            "in" : "query",
            "type" : "number"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "resultCode": {
                  "type": "number",
                  "example": 200
                },
                "resultMsg": {
                  "type": "string",
                  "example": "프로젝트 조회 완료"
                },
                "totalPage" : {
                  "type" : "number",
                  "example" : 1
                },
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "prj_id" : {
                        "type" : "number",
                        "example" : 15
                      },
                      "prj_name" : {
                        "type" : "string",
                        "example" : "프로젝트입니다"
                      },
                      "prj_description" : {
                        "type" : "string",
                        "example" : "v0.0.1"
                      },
                      "prj_lnk" : {
                        "type" : "string",
                        "example" : "https://example.com/"
                      },
                      "version_number" : {
                        "type" : "string",
                        "example" : "v0.0.1"
                      },
                      "step_number" : {
                        "type" : "number",
                        "example" : 1
                      },
                      "step_status" : {
                        "type" : "string",
                        "example" : "검토중"
                      },
                      "rgst_dtm" : {
                        "type" : "string",
                        "example" : "2024-08-26 16:01:40"
                      },
                      "updt_dtm" : {
                        "type" : "string",
                        "example" : "2024-08-26 16:01:40"
                      }, 
                      "dev_ids" : {
                        "type" : "string",
                        "example" : "13, 12"
                      },
                      "dev_user_name" : {
                        "type" : "string",
                        "example" : "테스터13, 테스터12"
                      },
                      "sec_ids" : {
                        "type" : "string",
                        "example" : "3"
                      },
                      "sec_user_name" : {
                        "type" : "string",
                        "example" : "테스터"
                      }
                    } 
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        },
        "security": [
          {
            "apiKeyAuth": [
              {
                "token": ""
              }
            ]
          }
        ]
      }
    },
...
  }
}